Implement complete CRUD functionality for AdminStudent
- Added Add, List, Edit, Delete actions in AdminStudentController
- Integrated Add.cshtml, List.cshtml, Edit.cshtml views
- Added Bootstrap modal for delete confirmation in Edit view
- Connected database using StudentDbContext
- Ensured redirect after each operation and proper form handling

+++ Migrations/StudentDbContextModelSnapshot.cs
@@ -35,8 +35,8 @@ namespace ResultProcessingSystem.Migrations
                     b.Property<string>("Name")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("Registration")
-                        .HasColumnType("int");
+                    b.Property<long>("Registration")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Session")
                         .HasColumnType("nvarchar(max)");
